requirements.txt
certifi==2018.4.16
chardet==3.0.4
dataclasses==0.6
Django==2.0.7
djongo==1.2.29
idna==2.7
lxml==4.2.3
pymongo==3.7.0
python-docx==0.8.6
pytz==2018.5
requests==2.19.1
sqlparse==0.2.4
urllib3==1.23
